merge

+++ ePLUS/Components/HeartRating.swift
@@ -9,7 +9,7 @@ import SwiftUI
 
 struct HeartRating: View {
     let heartSize: CGFloat = 20
-    @Binding var rating: Int
+    @Binding var rating: Float
     @State var status: [String] = ["hearthollow", "hearthollow", "hearthollow", "hearthollow", "hearthollow"]
     @State var changable: Bool = false
 
@@ -23,7 +23,7 @@ struct HeartRating: View {
                     .frame(width: heartSize, height: heartSize)
                     .onTapGesture {
                         if (changable) {
-                            updateRating(i+1)
+                            updateRating(Float(i+1))
                         }
                     }
             }
@@ -35,19 +35,19 @@ struct HeartRating: View {
         })
     }
     
-    func updateRating(_ rating: Int) {
+    func updateRating(_ rating: Float) {
         self.rating = rating
-        for i in 0..<self.rating {
+        for i in 0..<Int(self.rating) {
             self.status[i] = "heartfill"
         }
-        for i in self.rating..<5 {
+        for i in Int(self.rating)..<5 {
             self.status[i] = "hearthollow"
         }
     }
 }
 
 struct HeartRating_Previews_Container: View {
-    @State var rating: Int = 4
+    @State var rating: Float = 4.0
     
     var body: some View {
         HeartRating(rating: $rating)
